requirements.txt
paddlepaddle-gpu==2.6.1
paddleocr==2.8.1
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
websockets 